Constraints y ortografía en varias vistas. Corrección Vista Gráfica.

+++ SaludOS-Movil/Modelos/Grafica.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Grafica: Identifiable {
     var id: String
     var fechaYHoraToma: Date
-    var presionDiastolica: Int
-    var presionSistolica: Int
-    var pulso: Int
+    var presionDiastolica: Double
+    var presionSistolica: Double
+    var pulso: Double
 }
